requirements.txt
async-timeout==4.0.3
autopep8==2.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
idna==3.6
influxdb-client==1.31.0
jwt==1.3.1
paho-mqtt==1.6.1
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==4.6.0
requests==2.31.0
Rx==3.2.0
simple-watchdog-timer==0.1.1
six==1.16.0
tb-rest-client==3.6.2
tomli==2.0.1
urllib3==1.26.11
